+++ tests/packerlicious/test_provisioner_chef.py
@@ -1,0 +1,11 @@
+import pytest
+
+import packerlicious.provisioner as provisioner
+
+
+class TestChefSoloProvisioner(object):
+
+    def test_no_required_fields(self):
+        b = provisioner.ChefSolo()
+
+        b.to_dict()